fix: float had 64 bits length in migrations and models when 54 was max

+++ src/db/migrations/20200814180140-create-truck.js
@@ -16,7 +16,7 @@ module.exports = {
         type: Sequelize.TEXT
       },
       rate: {
-        type: Sequelize.FLOAT(64)
+        type: Sequelize.FLOAT
       },
       ownedByCompany: {
         type: Sequelize.BOOLEAN,
